Allow for searching for non-tex format files
Previously, we always defaulted to using the kpse_tex_file_format when
searching for files, which prevented using the library to search for
non-tex files (like font metrics, font glyphs, etc).

This adds a new Kpaths::guess_format_from_filename method which
guesses what format a filename has based on its extension. This is
based on a similar function in kpsewhich. Using this allows for almost
any format to be found.