Fix bug in `guess_format_from_filename`
It turns out that there was a bug in how I was initializing the
formats when they weren't already initialized. I was fetching a *copy*
of the format info struct, and then initializing the format if it
wasn't filled in, but this didn't end up filling in the copy of the
struct, so lookups would fail. Miraculously, in the test for this
function, I chose to look for a `.tex` file first, which then went and
initialized all of the formats, returned the default of
`kpse_tex_format`, so that lookup worked, and then in the later
lookups the formats were already initialized. Oops.

This fixes the problem by fetching a reference to the format info
struct, which then actually gets updated when `kpathsea_init_format`
runs.

Test plan: `cargo test -- --test-threads=1`