Implement Drop and Send for Kpaths
This adds a few useful traits for the Kpaths struct. I implemented
`Drop`, which is the destructor trait. For this, I'm just calling the
`kpathsea_finish` function to help free some memory.

I also added the trivial implementation of the `Send` trait. Since
kpathsea owns all of the data that is referenced by the kpathsea
pointer, this should be safe.

Test plan: `cargo test -- --test-threads=1`